Create UI Class [id=66302]
Found another small bug in the Interface implementation:
If you set the algorithm again, to the same algorithm, the function does
nothing, since there is no need to update with the same algorithm again.
There is the case when you load a different input matrix. You necesarily
need to set the algorithm again to point to the new input matrix, wether
it is the same alrogrithm or not. So, I removed the if statement to no
longer skip if trying to set same algorithm.